requirements.txt
selenium~=4.34.2
PyYAML~=6.0.2